requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pymongo>=4.6.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

